requirements.txt
python-nmap~=0.7.1
pydantic~=2.6.0
superagi-tools~=1.0.8